Increased the channel mask size to 64-bit and changed the masking logic

+++ include/sdr_pmr446.h
@@ -33,7 +33,7 @@ struct arguments
     float squelch_level;
     size_t waterfall;
     bool lowpass;
-    uint16_t channel_mask;
+    uint64_t channel_mask;
     lock_mode_e lock_mode;
 };
